pyproject.toml
[tool.poetry]
name = "ved"
version = "0.0.0"
description = "Python library dedicated to providing tools for common video problems."
authors = ["Raging Tiger <ragingtigertech@gmail.com>"]
license = "Proprietary"
readme = "README.md"
homepage = "https://github.com/RagingTiger/ved"
repository = "https://github.com/RagingTiger/ved"
packages = [
    { include = "ved", from = "src" },
]
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.urls]
Changelog = "https://github.com/RagingTiger/ved/releases"

[tool.poetry.dependencies]
python = "^3.9"
basc-py4chan = ">=0.6.6"
beautifulsoup4 = ">=4.12.2"
click = ">=8.0.1"
moviepy = ">=1.0.3"
Pillow = ">=10.0.1"
pathvalidate = ">=3.2.0"
requests = ">=2.31.0"
tqdm = ">=4.66.1"
youtube-dl = ">=2021.12.17"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
nox = ">=2023.4.22"
nox-poetry = ">=1.0.3"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
types-beautifulsoup4 = ">=4.12.0.7"
types-requests = ">=2.31.0.10"
types-tqdm = ">=4.66.0.3"

[tool.poetry.scripts]
ved = "ved.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["ved", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[[tool.mypy.overrides]]
module = ["basc_py4chan.*", "moviepy.*", "PIL.*", "youtube_dl.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
